GH-296 preserve Dave leaderboard decimal imported race totals in participant comparison and prevent int fallback to legacy values

+++ src/F1.Api/Dtos/AdminMigrationRunDtos.cs
@@ -147,11 +147,11 @@ public sealed record AdminMigrationPreseasonReasonCategorySummaryDto(
 
 public sealed record AdminMigrationParticipantComponentDeltaDto(
     string Subject,
-    int ImportedRacePoints,
+    decimal ImportedRacePoints,
     decimal CalculatedRacePoints,
     int ImportedPreseasonPoints,
     int CalculatedPreseasonPoints,
-    int ImportedTotalPoints,
+    decimal ImportedTotalPoints,
     decimal CalculatedTotalPoints,
     decimal NetDeltaPoints,
     string? TopReasonCode,
